+++ ceammc/ext/doc/docs.cmake
@@ -320,12 +320,13 @@ set(DOC_FILES
     hw.arduino
     hw.display
     hw.gamepad
-    hw.gpio
-    hw.gpio.rotenc
     hw.kbd_light
     hw.motu.avb
     hw.printer
-    hw.sensor.sr04
+    hw.rpi.gpio
+    hw.rpi.pwm
+    hw.rpi.rotenc
+    hw.rpi.sr04
     hw.serial
     hw.spi.max7219
     is_any