+++ ceammc/ext/src/hw/hw_rpi_dht11.cpp
@@ -1,10 +1,10 @@
-#include "hw_sensors_dht11.h"
+#include "hw_rpi_dht11.h"
 #include "args/argcheck.h"
 #include "ceammc_containers.h"
 #include "ceammc_factory.h"
 
-HwSensorDht11::HwSensorDht11(const PdArgs& args)
-    : HwDht11Base(args)
+HwRpiDht11::HwRpiDht11(const PdArgs& args)
+    : DispatchedObject<BaseObject>(args)
     , pin_(nullptr)
 {
     pin_ = new IntProperty("@pin", 4, PropValueAccess::INITONLY);
@@ -15,24 +15,24 @@ HwSensorDht11::HwSensorDht11(const PdArgs& args)
     createOutlet();
 }
 
-HwSensorDht11::~HwSensorDht11()
+HwRpiDht11::~HwRpiDht11()
 {
     ceammc_hw_gpio_dht11_free(dht_);
 }
 
-void HwSensorDht11::initDone()
+void HwRpiDht11::initDone()
 {
     dht_ = ceammc_hw_gpio_dht11_new(pin_->value(),
         { size_t(subscriberId()), [](size_t id) { Dispatcher::instance().send(NotifyMessage { id, 0 }); } }, //
         { this, [](void* data, const char* msg) {
-             auto obj = static_cast<HwSensorDht11*>(data);
+             auto obj = static_cast<HwRpiDht11*>(data);
              if (!obj)
                  return;
 
              Error(obj) << msg;
          } },
         { this, [](void* user, double temp, double hum) {
-             auto obj = static_cast<HwSensorDht11*>(user);
+             auto obj = static_cast<HwRpiDht11*>(user);
              if (!obj)
                  return;
 
@@ -43,12 +43,12 @@ void HwSensorDht11::initDone()
          } });
 }
 
-void HwSensorDht11::onBang()
+void HwRpiDht11::onBang()
 {
     ceammc_hw_gpio_dht11_measure(dht_);
 }
 
-void HwSensorDht11::m_poll(t_symbol* s, const AtomListView& lv)
+void HwRpiDht11::m_poll(t_symbol* s, const AtomListView& lv)
 {
     static const args::ArgChecker args("STATE:b");
     if (!args.check(lv, this))
@@ -57,13 +57,13 @@ void HwSensorDht11::m_poll(t_symbol* s, const AtomListView& lv)
     ceammc_hw_gpio_dht11_poll(dht_, lv.boolAt(0, false));
 }
 
-bool HwSensorDht11::notify(int /*code*/)
+bool HwRpiDht11::notify(int /*code*/)
 {
     return ceammc_hw_gpio_dht11_process(dht_);
 }
 
-void setup_hw_sensor_dht11()
+void setup_hw_rpi_dht11()
 {
-    ObjectFactory<HwSensorDht11> obj("hw.sensor.dht11");
-    obj.addMethod("poll", &HwSensorDht11::m_poll);
+    ObjectFactory<HwRpiDht11> obj("hw.rpi.dht11");
+    obj.addMethod("poll", &HwRpiDht11::m_poll);
 }