+++ ceammc/ext/src/hw/hw_rpi_dht11.h
@@ -1,20 +1,18 @@
-#ifndef HW_SENSORS_DHT11_H
-#define HW_SENSORS_DHT11_H
+#ifndef HW_RPI_DHT11_H
+#define HW_RPI_DHT11_H
 
 #include "ceammc_object.h"
 #include "ceammc_poll_dispatcher.h"
 #include "hw_rust.hpp"
 using namespace ceammc;
 
-using HwDht11Base = DispatchedObject<BaseObject>;
-
-class HwSensorDht11 : public HwDht11Base {
+class HwRpiDht11 : public DispatchedObject<BaseObject> {
     ceammc_hw_gpio_dht11* dht_ { nullptr };
     IntProperty* pin_ { nullptr };
 
 public:
-    HwSensorDht11(const PdArgs& args);
-    ~HwSensorDht11();
+    explicit HwRpiDht11(const PdArgs& args);
+    ~HwRpiDht11();
 
     void initDone() final;
     bool notify(int code) final;
@@ -23,6 +21,6 @@ public:
     void m_poll(t_symbol* s, const AtomListView& lv);
 };
 
-void setup_hw_sensor_dht11();
+void setup_hw_rpi_dht11();
 
-#endif // HW_SENSORS_DHT11_H
+#endif // HW_RPI_DHT11_H