+++ ceammc/ext/src/hw/hw_rpi_sr04.h
@@ -1,22 +1,20 @@
-#ifndef HW_SENSOR_SR04_H
-#define HW_SENSOR_SR04_H
+#ifndef HW_RPI_SR04_H
+#define HW_RPI_SR04_H
 
 #include "ceammc_object.h"
 #include "ceammc_poll_dispatcher.h"
 #include "hw_rust.hpp"
 using namespace ceammc;
 
-using HwSR04Base = DispatchedObject<BaseObject>;
-
-class HwSensorSR04 : public HwSR04Base {
+class HwRpiSr04 : public DispatchedObject<BaseObject> {
     ceammc_hw_gpio_sr04* sr04_ { nullptr };
     IntProperty* trigger_pin_ { nullptr };
     IntProperty* echo_pin_ { nullptr };
     IntProperty* poll_interval_ { nullptr };
 
 public:
-    HwSensorSR04(const PdArgs& args);
-    ~HwSensorSR04();
+    explicit HwRpiSr04(const PdArgs& args);
+    ~HwRpiSr04();
 
     void initDone() final;
     bool notify(int code) final;
@@ -25,6 +23,6 @@ public:
     void m_poll(t_symbol* s, const AtomListView& lv);
 };
 
-void setup_hw_sensor_sr04();
+void setup_hw_rpi_sr04();
 
-#endif // HW_SENSOR_SR04_H
+#endif // HW_RPI_SR04_H